refactor: re-organize components (zendro vs generic)
