--- src/components/fields/index.ts
@@ -1,3 +1,0 @@
-import StyledSelect from './select';
-
-export { StyledSelect };
