+++ src/components/inputs/index.ts
@@ -1,0 +1,8 @@
+import InputActions from './input-actions';
+import SelectInput from './select-input';
+import TextInput from './text-input';
+
+export * from './input-actions';
+export * from './text-input';
+
+export { InputActions, SelectInput, TextInput };
