+++ src/zendro/record-fields/float-field.tsx
@@ -1,9 +1,9 @@
 import { ReactElement } from 'react';
 import { Overwrite } from 'utility-types';
-import TextField, { TextFieldProps } from './text-field';
+import { TextInput, TextInputProps } from '@/components/inputs';
 
 type FloatFieldProps = Overwrite<
-  TextFieldProps,
+  TextInputProps,
   {
     onChange?: (value: number | null) => void;
     onError?: (value?: string) => void;
@@ -43,7 +43,7 @@ export default function FloatField({
   };
 
   return (
-    <TextField
+    <TextInput
       {...props}
       onChange={handleOnChange}
       onKeyDown={handleOnKeyDown}
