+++ src/zendro/record-fields/index.ts
@@ -1,3 +1,4 @@
+import ArrayField from './array-field';
 import BoolField from './bool-field';
 import DateTimeField from './datetime-field';
 import FloatField from './float-field';
@@ -5,17 +6,12 @@ import IntField from './int-field';
 import StringField from './string-field';
 
 import InputField from './input-field';
-import FieldIcons from './field-icons';
-import TextField from './text-field';
-import ArrayField from './array-field';
 
 export * from './bool-field';
 export * from './datetime-field';
 export * from './float-field';
-export * from './field-icons';
 export * from './int-field';
 export * from './string-field';
-export * from './text-field';
 export * from './array-field';
 export * from './input-field';
 
@@ -24,9 +20,7 @@ export {
   BoolField,
   DateTimeField,
   FloatField,
-  FieldIcons,
   InputField,
   IntField,
   StringField,
-  TextField,
 };
